+++ src/main/java/user_creator/UserCreationScreen.java
@@ -55,7 +55,7 @@ public class UserCreationScreen extends JFrame implements ActionListener {
     public void actionPerformed(ActionEvent evt) {
         try {
             userCreationController.accountCreator(username.getText(), password.getText(), repeatPassword.getText(),
-                    Float.parseFloat(newWeight), Float.parseFloat(newHeight), sex.getText(), birthday.getText());
+                    Double.parseDouble(newWeight), Double.parseDouble(newHeight), sex.getText(), birthday.getText());
             JOptionPane.showMessageDialog(this, "Account was created!");
         } catch (Exception e) {
             JOptionPane.showMessageDialog(this, "Accound creation failed!");
